// containers/web-server/frontend/src/lib/api-client/models/ReturnPeriodDamage.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ReturnPeriodDamage = {
    exposure: number;
    damage_amin: number;
    damage_mean: number;
    damage_amax: number;
    loss_amin: number;
    loss_mean: number;
    loss_amax: number;
    hazard: string;
    rcp: string;
    epoch: string;
    rp: number;
};